csc-assignment1/ExtraTask/requirements.txt
Django==2.0
django-webpack-loader==0.7.0
pip==20.1
pytz==2020.1
setuptools==41.2.0
sqlparse==0.2.4
bson==0.5.8
dataclasses==0.6
djangorestframework==3.11.0
djongo==1.2.30
idna==2.9
mongoengine==0.20.0
Pillow==7.1.2
requests==2.23.0
pymongo==3.10.1
python-dateutil==2.8.1
six==1.15.0
dnspython==1.16.0
certifi==2020.6.20
chardet==3.0.4
urllib3==1.25.9
absl-py==0.9.0
astunparse==1.6.3
cachetools==4.1.0
cycler==0.10.0
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
importlib-metadata==1.6.1
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.2
numpy==1.19.0
oauthlib==3.1.0
opencv-contrib-python==4.2.0.34
opt-einsum==3.2.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
Werkzeug==1.0.1
wheel==0.34.2
wrapt==1.12.1
zipp==3.1.0
slack==0.0.2
